06-workshops/01-AgentCore-runtime/01-hosting-agent/06-strands-with-skills/requirements.txt
strands-agents
strands-agents-tools
uv
boto3
bedrock-agentcore<=0.1.5
bedrock-agentcore-starter-toolkit==0.1.14
